+++ fn.py
@@ -578,41 +578,48 @@ def is_minimap_viewport(context=None):
 
     return True
 
-def get_headers_bottom_width(context, overlap=False) -> int:
+def get_header_margin(context, bottom=True, overlap=False) -> int:
     '''Return margin of bottom aligned headers
     (Possible regions: HEADER, TOOL_HEADER, ASSET_SHELF, ASSET_SHELF_HEADER)
     '''
     if not context.preferences.system.use_region_overlap:
         return 0
+    
+    if bottom:
+        side_name = 'BOTTOM'
+    else:
+        side_name = 'TOP'
 
-    bottom_margin = 0
+    margin = 0
     ## asset shelf header should be only added only if shelf deployed
     regions = context.area.regions
     header = next((r for r in regions if r.type == 'HEADER'), None)
     tool_header = next((r for r in regions if r.type == 'TOOL_HEADER'), None)
     asset_shelf = next((r for r in regions if r.type == 'ASSET_SHELF'), None)
 
-    if header.alignment == 'BOTTOM':
-        bottom_margin += header.height
-    if tool_header.alignment == 'BOTTOM':
-        bottom_margin += tool_header.height
+    if header.alignment == side_name : 
+        margin += header.height
+
+    if tool_header.alignment == side_name : 
+        margin += tool_header.height
     
-    if asset_shelf.height > 1:
+    ## calculate asset height only if checking bottom width
+    if bottom and asset_shelf.height > 1:
         
         if not overlap:
             ## Header of asset shelf should only be added if shelf open
             asset_shelf_header = next((r for r in regions if r.type == 'ASSET_SHELF_HEADER'), None)
-            bottom_margin += asset_shelf.height + asset_shelf_header.height
+            margin += asset_shelf.height + asset_shelf_header.height
         else:
             ## Only if toggle icon is centered
-            bottom_margin += asset_shelf.height
+            margin += asset_shelf.height
 
     # for r in context.area.regions:
     #     ## 'ASSET_SHELF_HEADER' is counted even when not visible
     #     if r.alignment == 'BOTTOM' and r.type != 'ASSET_SHELF_HEADER':
-    #         bottom_margin += r.height
+    #         margin += r.height
     
-    return bottom_margin
+    return margin
 
 def set_gizmo_settings(gz, icon,
         color=(0.0, 0.0, 0.0),