+++ tool_presets/__init__.py
@@ -1,5 +1,6 @@
-from . import (properties,
-               tool_list,
+from .. import (
+            properties,
+            tool_list,
                )
 
 modules = (